backend/requirements.txt
tensorflow==2.21.0
keras==3.14.0
numpy==1.26.4
opencv-python==4.9.0.80
flask==3.0.0
flask-cors==4.0.0
pillow==10.2.0